Added functions to set steps per revolution
Upgraded motor and main board firmware such that steps per
revolution for the motor board's stepper motors can be set
through UDP. The main board receives a MSG_SET_STEPS_PER_REV
message and forwards it through I2C to the motor board. The
JSON encoded UDP message to the main board needs to contain
a field "steps" and a value of 0 < steps < 80.